pyproject.toml
[project]
name = "label-software"
version = "0.1.0"
description = "Video annotation and labeling tool for fall detection research"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.39",
    "opencv-python>=4.11.0",
    "pillow>=11.0.0",
    "watchdog>=6.0.0",
    "httpx>=0.28.0",
    "pydantic-settings>=2.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "google-auth-httplib2>=0.2.0",
    "google-api-python-client>=2.0.0",
    "werkzeug>=3.0.0",
]

[project.scripts]
label-dev = "label_software.cli:dev"
label-serve = "label_software.cli:serve"
label-build = "label_software.cli:build"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/label_software"]

[dependency-groups]
dev = ["pytest>=8.3.5", "httpx>=0.28.0", "ruff>=0.9.0"]
